requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.0
sqlalchemy==2.0.23
pandas==1.5.3
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1
aiofiles==23.2.1
python-multipart==0.0.6
